Fiddle with asciidoc reporting
Trying out a scheme with sections so that a toc is automatically generated by
asciidoc as summary.  Also presenting public details in tables, but not sure
if this is the final form. It does add visual weight to a section which is nice.

General fixes while doing above.

Not happy with asciidoc generation code. It is doing too much work dealing
with deep-diff format and figuring out change inheritance in diff tree.